// core/packages/client/src/pages/gen/tnoodle_i18n.ts
export type TnoodleLocale = "br" | "da" | "de" | "en" | "eo" | "es" | "et" | "fi" | "fr" | "hr" | "hu" | "id" | "it" | "ja" | "ko" | "nb" | "pl" | "pt" | "pt-BR" | "rm" | "ro" | "ru" | "sl" | "vi" | "zh-CN" | "zh-TW";

export const TNOODLE_LOCALES: readonly TnoodleLocale[] = ["br","da","de","en","eo","es","et","fi","fr","hr","hu","id","it","ja","ko","nb","pl","pt","pt-BR","rm","ro","ru","sl","vi","zh-CN","zh-TW"] as const;

export interface TnoodleTranslations {
  fmc: {
    scrambleXofY: string;
    scramble: string;
    round: string;
    attempt: string;
    competitor: string;
    warning: string;
    event: string;
    graded: string;
    result: string;
    competition: string;
    registrantId: string;
    rule1: string;
    rule2: string;
    rule3: string;
    rule4: string;
    rule5: string;
    rule6: string;
    faceMoves: string;
    rotations: string;
    clockwise: string;
    counterClockwise: string;
    double: string;
    scrambleOnSeparateSheet: string;
  };
}

export const TNOODLE_I18N: Record<TnoodleLocale, TnoodleTranslations> = {
  "br": {
    "fmc": {
      "scrambleXofY": "Meskad %{scrambleIndex} diwar %{scrambleCount}",
      "scramble": "Meskad",
      "round": "Tro",
      "attempt": "Taol-esae",
      "competitor": "Kenstriver",
      "warning": "MARD OC'H AR C'HENSTRIVER, LAOSKIT GOULLO.",
      "event": "Diluziadur Gwellekaet",
      "graded": "Gwiriet gant",
      "result": "Disoc'h",
      "competition": "Kenstrivadeg",
      "registrantId": "Anaouder an enskrivadur",
      "rule1": "60 munutenn ho peus evit kavout ha skrivañ un diskoulm.",
      "rule2": "Skrivit 1 fiñvadenn dre tired. Evit dilemel ur fiñvadenn, duit anezhi en un doare sklaer.",
      "rule3": "N'hall ket bezañ deveret ho tiskoulm war-eeun diwar lodenn-pe-lodenn eus ar meskad.",
      "rule4": "Rankout a ra ho tiskoulm kaout d'ar muiañ %{maxMoves} a fiñvadennoù, troadur hag all.",
      "rule5": "Kontet e vo ho tisoc'h en OBTM.",
      "rule6": "Implijit notadur ar Mellad 12 Reoliadur eus ar WCA hepken. Ma n'oc'h ket sur, grit gant ar fiñvadennoù-mañ hepken:",
      "faceMoves": "Fiñvadennoù ar fasoù",
      "rotations": "Troiadurioù",
      "clockwise": "A-du gant an horolaj",
      "counterClockwise": "En amgin d'an horolaj",
      "double": "Doubl",
      "scrambleOnSeparateSheet": "Trugarez da sellet ouzh ar follenn dispartiet evit ar meskad."
    }
  },
// ...
};